+++ src/core/models/titrations/1_1_Model.h
@@ -55,7 +55,7 @@ public:
 
     virtual inline QString SpeciesName(int i) const override
     {
-        if (i == 1)
+        if (i == 0)
             return AB;
         else
             return QString();